new checkout request model and routes working. also extracted inventory routes into their own folder
