+++ api/routes/bookRoutes.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const db = require("../../data/dbConfig");
 const router = express.Router();
-// const Books = require("../helpers/bookModel");
-const CheckedOut = require("../helpers/checkedOutModel");
+const Books = require("../helpers/bookModel");
 
 const rp = require("request-promise");
 const { parseString } = require("xml2js");
@@ -33,7 +32,7 @@ router.post("/search", (req, res) => {
 
 router.get("/", async (req, res) => {
   try {
-    const books = await CheckedOut.getAllBooks().orderBy("bookId");
+    const books = await Books.getAllBooks().orderBy("bookId");
     res.status(200).json(books);
   } catch (error) {
     res.status(500).json(error);
@@ -45,14 +44,11 @@ router.get("/", async (req, res) => {
 router.post("/", async (req, res) => {
   try {
     const book = await db("books").insert(req.body);
-    const newBookList = await db("books");
-    const newBook = await db("books")
-      .where({ title: req.body.title })
-      .first();
+    // const newBook = await db("books")
+    //   .where({ title: req.body.title })
+    //   .first();
     if (book) {
-      return res
-        .status(200)
-        .json({ message: "Book successfully added", newBook });
+      return res.status(200).json({ message: "Book successfully added" });
     } else {
       return res.status(404).json(error);
     }
@@ -65,7 +61,7 @@ router.post("/", async (req, res) => {
 
 router.get("/:bookId", async (req, res) => {
   try {
-    const book = await CheckedOut.getBookById(req.params.bookId).first();
+    const book = await Books.getBookById(req.params.bookId).first();
     if (book) {
       res.status(200).json(book);
     } else {
